Added files via upload
These are my Collective Methods Web Vandalism files. I was a bit intimidated by this assignment at the start because I don't have much coding experience. But it became a lot of fun as I tried it out. My first try was pretty lame but I kept it to exhibit my first attempt. (What has been seen...) My second shot seemed a little better to me, as with that one and number four I experimented with how to make it a little bit more subtle.  With number three, the NYUSH page, I experimented with repetition and actually with the text a little bit. It helped to play around a little with the code. Number 5 (Trump/Pham 2016) was my personal favorite, and to me signified my conquest of inserting new pictures into code! Excited to learn more.